moved manager perms, moving to destroying user and correcting permissions validation
