+++ src/main/java/com/group16b/DomainLayer/ProductionCompany/membership/HierarchyNodeData.java
@@ -2,7 +2,7 @@ package com.group16b.DomainLayer.ProductionCompany.membership;
 
 import java.util.Set;
 
-import com.group16b.DomainLayer.User.Roles.ManagerPermissions;
+import com.group16b.DomainLayer.ProductionCompany.membership.ManagerPermissions;
 import com.group16b.DomainLayer.ProductionCompany.membership.RoleType;
 
 public class HierarchyNodeData {
